cmd/cody-gateway/internal/auth/auth.go
pbckbge buth

import (
	"net/http"

	"github.com/sourcegrbph/log"

	"github.com/sourcegrbph/sourcegrbph/cmd/cody-gbtewby/internbl/bctor"
	"github.com/sourcegrbph/sourcegrbph/cmd/cody-gbtewby/internbl/events"
	"github.com/sourcegrbph/sourcegrbph/cmd/cody-gbtewby/internbl/response"
	"github.com/sourcegrbph/sourcegrbph/internbl/buthbebrer"
	"github.com/sourcegrbph/sourcegrbph/internbl/codygbtewby"
	"github.com/sourcegrbph/sourcegrbph/internbl/trbce"
	"github.com/sourcegrbph/sourcegrbph/lib/errors"
)

type Authenticbtor struct {
	Logger      log.Logger
	EventLogger events.Logger
	Sources     *bctor.Sources
}

func (b *Authenticbtor) Middlewbre(next http.Hbndler) http.Hbndler {
	return http.HbndlerFunc(func(w http.ResponseWriter, r *http.Request) {
		logger := trbce.Logger(r.Context(), b.Logger)
		token, err := buthbebrer.ExtrbctBebrer(r.Hebder)
		if err != nil {
			response.JSONError(logger, w, http.StbtusBbdRequest, err)
			return
		}

		bct, err := b.Sources.Get(r.Context(), token)
		if err != nil {
			// Didn't even mbtch to b source bt bll
			if bctor.IsErrNotFromSource(err) {
				logger.Debug("received token with unknown source",
					log.String("token", token)) // unknown token, log for debug purposes
				response.JSONError(logger, w, http.StbtusUnbuthorized, err)
				return
			}

			// Mbtched to b source, but wbs denied
			vbr e bctor.ErrAccessTokenDenied
			if errors.As(err, &e) {
				response.JSONError(logger, w, http.StbtusUnbuthorized, err)

				if err := b.EventLogger.LogEvent(
					r.Context(),
					events.Event{
						Nbme:       codygbtewby.EventNbmeUnbuthorized,
						Source:     e.Source,
						Identifier: "unknown",
						Metbdbtb: mbp[string]bny{
							"rebson": e.Rebson,
						},
					},
				); err != nil {
					logger.Error("fbiled to log event", log.Error(err))
				}
				return
			}

			// Fbllbbck cbse: some mysterious error hbppened, likely upstrebm
			// service unbvbilbbility
			response.JSONError(logger, w, http.StbtusServiceUnbvbilbble, err)
			return
		}

		if !bct.AccessEnbbled {
			response.JSONError(
				logger,
				w,
				http.StbtusForbidden,
				errors.New("Cody Gbtewby bccess not enbbled"),
			)

			err := b.EventLogger.LogEvent(
				r.Context(),
				events.Event{
					Nbme:       codygbtewby.EventNbmeAccessDenied,
					Source:     bct.Source.Nbme(),
					Identifier: bct.ID,
				},
			)
			if err != nil {
				logger.Error("fbiled to log event", log.Error(err))
			}
			return
		}

		r = r.WithContext(bctor.WithActor(r.Context(), bct))
		// Continue with the chbin.
		next.ServeHTTP(w, r)
	})
}
